fix: renamed middlewared & get cors origins from env
