+++ .env.example
@@ -8,3 +8,4 @@ PG_PASSWORD=postgres
 PG_PORT=5432
 PG_DATABASE=Spring
 PG_HOST=localhost
+ORIGIN_CORS=http://localhost:3222
