--- src/configs/swaggerOptions.ts
@@ -1,32 +1,0 @@
-import { SwaggerDefinition } from 'swagger-jsdoc';
-
-const swaggerDefinition: SwaggerDefinition = {
-  openapi: '3.0.0',
-  info: {
-    title: 'My API',
-    version: '1.0.0',
-    description: 'API for spring front',
-  },
-  servers: [
-    {
-      url: 'http://localhost:4444',
-      description: 'Development server',
-    },
-  ],
-  components: {
-    securitySchemes: {
-      bearerAuth: {
-        type: 'http',
-        scheme: 'bearer',
-        bearerFormat: 'JWT',
-      },
-    },
-  },
-};
-
-const options = {
-  swaggerDefinition,
-  apis: ['./src/modules/**/*.routes.ts'],
-};
-
-export default options;
